+++ client/pub-sockets.js
@@ -1,0 +1,31 @@
+/*
+ * pub-sockets.js
+ * browserify entry point
+ * connect/disconnect socket.io
+ *
+ * TODO: consolidate with socket.io script
+ * copyright 2015, Jurgen Leschner - github.com/jldec - MIT license
+*/
+
+var debug = require('debug')('pub:sockets');
+
+if (window.io) {
+
+  debug('socket:connect');
+  var socket = io();
+
+  socket.on('reload', function() {
+    // in-browser generator case, just notify
+    if (window.generator) {
+      return window.generator.emit('notify', 'save');
+    }
+    debug('socket:reload');
+    location.reload();
+  });
+
+  $(window).on('beforeunload', function() {
+    debug('socket:disconnect');
+    socket.disconnect();
+  })
+
+} else console.log('no socket.io')