+++ client/pub-ux.js
@@ -1,34 +1,9 @@
 /*
  * pub-ux.js
- * browserify entry point
- * connect/disconnect socket.io and inject pub UI into dom
+ * browserify entry point for legacy pub-pkg-editor
  *
- * TODO: fix /pub/ button logic to handle non-pub urls ending in /pub/
  * copyright 2015, Jurgen Leschner - github.com/jldec - MIT license
 */
-
-var debug = require('debug')('pub:ux');
-
-if (window.io) {
-
-  debug('socket:connect');
-  var socket = io();
-
-  socket.on('reload', function() {
-    // in-browser generator case, just notify
-    if (window.generator) {
-      return window.generator.emit('notify', 'save');
-    }
-    debug('socket:reload');
-    location.reload();
-  });
-
-  $(window).on('beforeunload', function() {
-    debug('socket:disconnect');
-    socket.disconnect();
-  })
-
-} else console.log('no socket.io')
 
 $(function(){